// src/components/request/types.ts
export interface UserInstance {
  id: number;
  avatar: string | null;
  id_document: string | null;
  proof_of_address: string | null;
  next_of_kin: string | null;
  bank_accounts: any[];
  role: string | null;
  team_status: string | null;
  team_company_name: string | null;
  in_review: boolean;
  expected_activities: boolean;
  trial_days: number;
  trial_display: boolean;
  is_owner: boolean;
  developer_link: string | null;
  last_login: string | null;
  username: string | null;
  bvn: string | null;
  year_of_birth: string | null;
  parents_ids: string | null;
  date_of_birth: string | null;
  email: string | null;
  first_name: string | null;
  lang: string | null;
  last_name: string | null;
  middle_name: string | null;
  phone: string | null;
  status: boolean;
  is_blocked: boolean;
  gender: string | null;
  default_timezone: string | null;
  address: string | null;
  yearly_income: string | null;
  employment_status: string | null;
  monthly_income: number | null;
  occupation: string | null;
  marital_status: string | null;
  company_name: string | null;
  company_address: string | null;
  cac_number: string | null;
  highest_education: string | null;
  referral_code: string | null;
  push_notification_id: string | null;
  enable_2fa: boolean;
  recovery_code: string | null;
  is_qualified_for_line_of_credit: boolean;
  is_qualified_for_mortgage: boolean;
  last_contact_update: string | null;
  push_notification_allowed: boolean;
  biometric_authentication: boolean;
  is_private: boolean;
  about_dev: string | null;
  initial_status: string | null;
  passward_change_time: string | null;
  referred_by: string | null;
  sign_up_time: string | null;
  dynamic_referral: string | null;
  is_tier3: boolean;
  use_security_question: boolean;
  security_question: string | null;
  security_answer: string | null;
  developer_feedback: string | null;
  referral_status: string | null;
  outreach: string | null;
  agent_id: string | null;
  invite_code: string | null;
  co_parent_invite: string | null;
  mfa_secret: string | null;
  is_mfa: boolean;
  last_activity: string | null;
  approved_at: string | null;
  country: number;
  annual_income: string | null;
  channel: string | null;
  groups: number[];
}

export interface AgentInstance {
  customer_info: string | null;
  id: number;
  bvn: string | null;
  agent_id: string | null;
  username: string | null;
  approved_at: string | null;
  date_of_birth: string | null;
  first_name: string | null;
  last_name: string | null;
  middle_name: string | null;
  user_status: string | null;
  phone: string | null;
  is_mfa: boolean;
  avatar: string | null;
  monthly_income: number | null;
  email: string | null;
  address: string | null;
  gender: string | null;
  occupation: string | null;
  marital_status: string | null;
  company_name: string | null;
  yearly_income: string | null;
  highest_education: string | null;
  employment_status: string | null;
  sign_up_time: string | null;
  company_address: string | null;
}
// ...
export interface CommissionRequestInstance {
  id: number;
  agent: AgentInstance;
  created_at: string;
  updated_at?: string | null;

  property_info: string | null;
  date_sold: string | null;
  customer: {
    id: number;
    user: UserInstance;
    developer: UserInstance;
    phone: string | null;
    sign_up_time: string | null;
    avatar: string | null;
    first_name: string | null;
    last_name: string | null;
    utility_bills: {
      utility_bill: string | null;
      uploaded_at: string | null;
      utility_bill_type: string | null;
      customer: number;
    }[];
    documents: {
      id: number;
      document: string | null;
      store: string | null;
      created_at: string;
      extended_user: number;
    }[];
    offset: string | null;
    customer_type: string | null;
    active: boolean;
    customer_ref: string | null;
    blacklisted: boolean;
    refunded_balance: string | null;
    utility_bill: string | null;
    utility_bill_updated_at: string | null;
    alternate_country_code: string | null;
    wallet_waitlist: boolean;
    is_tier3: boolean;
    gender: string | null;
    referred_at: string | null;
    created_at: string;
    password: string | null;
    default_timezone: string | null;
    highest_education: string | null;
    address: string | null;
    employment_status: string | null;
    monthly_income: number | null;
    yearly_income: string | null;
    occupation: string | null;
    marital_status: string | null;
    company_name: string | null;
    company_address: string | null;
    email: string | null;
    secondary_email: string | null;
    currency: string | null;
    middle_name: string | null;
    director_name: string | null;
    alternate_phone: string | null;
    company_phone: string | null;
    year_of_birth: string | null;
    bvn: string | null;
    email_verified: boolean;
    date_of_birth: string | null;
    outreach: string | null;
    username: string | null;
    extended_user: boolean;
    biometric_authentication: boolean;
    is_blocked: boolean;
    status: boolean;
    push_notification_allowed: boolean;
    stealth_mode: boolean;
    source_of_funds: string | null;
    nationality: string | null;
    country_of_residence: string | null;
    blacklist_date: string | null;
    business: number;
    referred_by: string | null;
    country: number;
    created_by: {
      id: number;
      first_name: string | null;
      last_name: string | null;
      email: string | null;
    } | null;
    blacklisted_by: {
      id: number;
      first_name: string | null;
      last_name: string | null;
      email: string | null;
    } | null;
  };
  status: string | null;
  request_type: string | null;
  response_note?: string | null;
}